+++ test/e2e/web-socket-server-and-url.test.js
@@ -420,90 +420,3 @@ for (const webSocketServerType of webSocketServerTypes) {
     });
   });
 }
-
-describe('Client console.log', () => {
-  const baseOptions = {
-    port: port2,
-    host: '0.0.0.0',
-  };
-  const webSocketServerTypesLog = [
-    {},
-    { webSocketServer: 'sockjs' },
-    { webSocketServer: 'ws' },
-  ];
-
-  const cases = [
-    {
-      title: 'hot disabled',
-      options: {
-        hot: false,
-      },
-    },
-    {
-      title: 'hot enabled',
-      options: {
-        hot: true,
-      },
-    },
-    {
-      title: 'liveReload disabled',
-      options: {
-        liveReload: false,
-      },
-    },
-    {
-      title: 'liveReload enabled',
-      options: {
-        liveReload: true,
-      },
-    },
-    {
-      title: 'liveReload & hot are disabled',
-      options: {
-        liveReload: false,
-        hot: false,
-      },
-    },
-    {
-      title: 'client logging is none',
-      options: {
-        client: {
-          logging: 'none',
-        },
-      },
-    },
-  ];
-
-  webSocketServerTypesLog.forEach(async (mode) => {
-    cases.forEach(async ({ title, options }) => {
-      title += ` (${
-        Object.keys(mode).length ? mode.webSocketServer : 'default'
-      })`;
-
-      options = { ...mode, ...options };
-
-      const testOptions = Object.assign({}, baseOptions, options);
-
-      it(title, (done) => {
-        testServer.startAwaitingCompilation(config, testOptions, async () => {
-          const res = [];
-          const { page, browser } = await runBrowser();
-
-          page.goto(`http://localhost:${port2}/main`);
-          page.on('console', ({ _text }) => {
-            res.push(_text);
-          });
-
-          // wait for load before closing the browser
-          await page.waitForNavigation({ waitUntil: 'load' });
-          await page.waitForTimeout(beforeBrowserCloseDelay);
-          await browser.close();
-
-          // Order doesn't matter, maybe we should improve that in future
-          await expect(res.sort()).toMatchSnapshot();
-          await testServer.close(done);
-        });
-      });
-    });
-  });
-});
