+++ test/server/open-option.test.js
@@ -35,7 +35,7 @@ describe('"open" option', () => {
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://localhost:${port}/`, {
           wait: false,
         });
 
@@ -58,7 +58,7 @@ describe('"open" option', () => {
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('https://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`https://localhost:${port}/`, {
           wait: false,
         });
 
@@ -72,16 +72,17 @@ describe('"open" option', () => {
 
   it("should work with '0.0.0.0' host", (done) => {
     const compiler = webpack(config);
+    const host = '0.0.0.0';
     const server = createServer(compiler, {
-      open: true,
+      host,
       port,
-      host: '0.0.0.0',
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://0.0.0.0:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           wait: false,
         });
 
@@ -90,21 +91,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, '0.0.0.0');
+    server.listen(port, host);
   });
 
   it("should work with '::' host", (done) => {
     const compiler = webpack(config);
+    const host = '::';
     const server = createServer(compiler, {
-      open: true,
+      host,
       port,
-      host: '::',
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://[::]:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://[${host}]:${port}/`, {
           wait: false,
         });
 
@@ -113,21 +115,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, '::');
+    server.listen(port, host);
   });
 
   it("should work with 'localhost' host", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: true,
+      host,
       port,
-      host: 'localhost',
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           wait: false,
         });
 
@@ -136,21 +139,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with '127.0.0.1' host", (done) => {
     const compiler = webpack(config);
+    const host = '127.0.0.1';
     const server = createServer(compiler, {
-      open: true,
+      host,
       port,
-      host: '127.0.0.1',
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://127.0.0.1:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           wait: false,
         });
 
@@ -159,21 +163,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, '127.0.0.1');
+    server.listen(port, host);
   });
 
   it("should work with '::1' host", (done) => {
     const compiler = webpack(config);
+    const host = '::1';
     const server = createServer(compiler, {
-      open: true,
+      host,
       port,
-      host: '::1',
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://[::1]:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://[${host}]:${port}/`, {
           wait: false,
         });
 
@@ -182,21 +187,21 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, '::1');
+    server.listen(port, host);
   });
 
   it(`should work with '${internalIPv4}' host`, (done) => {
     const compiler = webpack(config);
     const server = createServer(compiler, {
-      open: true,
-      port,
       host: internalIPv4,
+      port,
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith(`http://${internalIPv4}:8117/`, {
+        expect(open).toHaveBeenCalledWith(`http://${internalIPv4}:${port}/`, {
           wait: false,
         });
 
@@ -235,16 +240,17 @@ describe('"open" option', () => {
 
   it('should work with boolean', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: true,
+      host,
       port,
-      host: 'localhost',
+      open: true,
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           wait: false,
         });
 
@@ -253,15 +259,16 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with boolean but don't close with 'false' value", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: false,
+      host,
       port,
-      host: 'localhost',
+      open: false,
       static: false,
     });
 
@@ -274,21 +281,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with relative string', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: 'index.html',
+      host,
       port,
-      host: 'localhost',
+      open: 'index.html',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/index.html`, {
           wait: false,
         });
 
@@ -297,21 +305,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with relative string starting with "/"', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: '/index.html',
+      host,
       port,
-      host: 'localhost',
+      open: '/index.html',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/index.html`, {
           wait: false,
         });
 
@@ -320,13 +329,14 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with absolute string', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: 'http://localhost:8117/index.html',
+      open: `http://${host}:${port}/index.html`,
       port,
       host: 'localhost',
       static: false,
@@ -334,7 +344,7 @@ describe('"open" option', () => {
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/index.html`, {
           wait: false,
         });
 
@@ -343,15 +353,16 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with multiple relative strings', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: ['first.html', 'second.html'],
-      port,
       host: 'localhost',
+      port,
+      open: ['first.html', 'second.html'],
       static: false,
     });
 
@@ -359,14 +370,14 @@ describe('"open" option', () => {
       server.close(() => {
         expect(open).toHaveBeenNthCalledWith(
           1,
-          'http://localhost:8117/first.html',
+          `http://${host}:${port}/first.html`,
           {
             wait: false,
           }
         );
         expect(open).toHaveBeenNthCalledWith(
           2,
-          'http://localhost:8117/second.html',
+          `http://${host}:${port}/second.html`,
           {
             wait: false,
           }
@@ -377,18 +388,19 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with multiple absolute strings', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: [
-        'http://localhost:8117/first.html',
-        'http://localhost:8117/second.html',
-      ],
-      port,
       host: 'localhost',
+      port,
+      open: [
+        `http://${host}:${port}/first.html`,
+        `http://${host}:${port}/second.html`,
+      ],
       static: false,
     });
 
@@ -396,14 +408,14 @@ describe('"open" option', () => {
       server.close(() => {
         expect(open).toHaveBeenNthCalledWith(
           1,
-          'http://localhost:8117/first.html',
+          `http://${host}:${port}/first.html`,
           {
             wait: false,
           }
         );
         expect(open).toHaveBeenNthCalledWith(
           2,
-          'http://localhost:8117/second.html',
+          `http://${host}:${port}/second.html`,
           {
             wait: false,
           }
@@ -414,21 +426,22 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with empty object', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
-      open: {},
+      host,
       port,
-      host: 'localhost',
+      open: {},
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           wait: false,
         });
 
@@ -437,23 +450,24 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object and with the boolean value of 'target' option", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         target: true,
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           wait: false,
         });
 
@@ -462,23 +476,24 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object and with the 'target' option", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         target: 'index.html',
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/index.html`, {
           wait: false,
         });
 
@@ -487,17 +502,18 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object and with multiple values of the 'target' option", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         target: ['first.html', 'second.html'],
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
@@ -505,14 +521,14 @@ describe('"open" option', () => {
       server.close(() => {
         expect(open).toHaveBeenNthCalledWith(
           1,
-          'http://localhost:8117/first.html',
+          `http://${host}:${port}/first.html`,
           {
             wait: false,
           }
         );
         expect(open).toHaveBeenNthCalledWith(
           2,
-          'http://localhost:8117/second.html',
+          `http://${host}:${port}/second.html`,
           {
             wait: false,
           }
@@ -523,23 +539,24 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object and with the 'app' option", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         app: 'google-chrome',
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           app: { name: 'google-chrome' },
           wait: false,
         });
@@ -549,23 +566,24 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object and with the 'app' and 'arguments' options", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         app: { name: 'google-chrome', arguments: ['--incognito'] },
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/`, {
           app: { name: 'google-chrome', arguments: ['--incognito'] },
           wait: false,
         });
@@ -575,24 +593,25 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it('should work with object with "target" and "app" options', (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         target: 'index.html',
         app: 'google-chrome',
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
+        expect(open).toHaveBeenCalledWith(`http://${host}:${port}/index.html`, {
           app: { name: 'google-chrome' },
           wait: false,
         });
@@ -602,18 +621,19 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object, with multiple value of the 'target' option and with the 'app' and 'arguments' options", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
         target: ['first.html', 'second.html'],
         app: { name: 'google-chrome', arguments: ['--incognito'] },
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
@@ -621,7 +641,7 @@ describe('"open" option', () => {
       server.close(() => {
         expect(open).toHaveBeenNthCalledWith(
           1,
-          'http://localhost:8117/first.html',
+          `http://${host}:${port}/first.html`,
           {
             wait: false,
             app: { name: 'google-chrome', arguments: ['--incognito'] },
@@ -629,7 +649,7 @@ describe('"open" option', () => {
         );
         expect(open).toHaveBeenNthCalledWith(
           2,
-          'http://localhost:8117/second.html',
+          `http://${host}:${port}/second.html`,
           {
             wait: false,
             app: { name: 'google-chrome', arguments: ['--incognito'] },
@@ -641,18 +661,19 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should work with object, with multiple value of the 'target' option (relative and absolute URLs) and with the 'app' option with arguments", (done) => {
     const compiler = webpack(config);
+    const host = 'localhost';
     const server = createServer(compiler, {
+      host,
+      port,
       open: {
-        target: ['first.html', 'http://localhost:8117/second.html'],
+        target: ['first.html', `http://${host}:${port}/second.html`],
         app: { name: 'google-chrome', arguments: ['--incognito'] },
       },
-      port,
-      host: 'localhost',
       static: false,
     });
 
@@ -660,7 +681,7 @@ describe('"open" option', () => {
       server.close(() => {
         expect(open).toHaveBeenNthCalledWith(
           1,
-          'http://localhost:8117/first.html',
+          `http://${host}:${port}/first.html`,
           {
             wait: false,
             app: { name: 'google-chrome', arguments: ['--incognito'] },
@@ -668,7 +689,7 @@ describe('"open" option', () => {
         );
         expect(open).toHaveBeenNthCalledWith(
           2,
-          'http://localhost:8117/second.html',
+          `http://${host}:${port}/second.html`,
           {
             wait: false,
             app: { name: 'google-chrome', arguments: ['--incognito'] },
@@ -680,7 +701,7 @@ describe('"open" option', () => {
     });
 
     compiler.run(() => {});
-    server.listen(port, 'localhost');
+    server.listen(port, host);
   });
 
   it("should log warning when can't open", (done) => {
@@ -688,20 +709,19 @@ describe('"open" option', () => {
 
     const compiler = webpack(config);
     const server = createServer(compiler, {
-      open: true,
       port,
-      host: 'localhost',
+      open: true,
       static: false,
     });
     const loggerWarnSpy = jest.spyOn(server.logger, 'warn');
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/', {
+        expect(open).toHaveBeenCalledWith(`http://localhost:${port}/`, {
           wait: false,
         });
         expect(loggerWarnSpy).toHaveBeenCalledWith(
-          'Unable to open "http://localhost:8117/" page. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".'
+          `Unable to open "http://localhost:${port}/" page. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".`
         );
 
         loggerWarnSpy.mockRestore();
@@ -726,11 +746,14 @@ describe('"open" option', () => {
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
-          wait: false,
-        });
+        expect(open).toHaveBeenCalledWith(
+          `http://localhost:${port}/index.html`,
+          {
+            wait: false,
+          }
+        );
         expect(loggerWarnSpy).toHaveBeenCalledWith(
-          'Unable to open "http://localhost:8117/index.html" page. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".'
+          `Unable to open "http://localhost:${port}/index.html" page. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".`
         );
 
         loggerWarnSpy.mockRestore();
@@ -758,12 +781,15 @@ describe('"open" option', () => {
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
-          app: { name: 'google-chrome' },
-          wait: false,
-        });
+        expect(open).toHaveBeenCalledWith(
+          `http://localhost:${port}/index.html`,
+          {
+            app: { name: 'google-chrome' },
+            wait: false,
+          }
+        );
         expect(loggerWarnSpy).toHaveBeenCalledWith(
-          'Unable to open "http://localhost:8117/index.html" page in "google-chrome" app. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".'
+          `Unable to open "http://localhost:${port}/index.html" page in "google-chrome" app. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".`
         );
 
         loggerWarnSpy.mockRestore();
@@ -794,15 +820,18 @@ describe('"open" option', () => {
 
     compiler.hooks.done.tap('webpack-dev-server', () => {
       server.close(() => {
-        expect(open).toHaveBeenCalledWith('http://localhost:8117/index.html', {
-          app: {
-            name: 'google-chrome',
-            arguments: ['--incognito', '--new-window'],
-          },
-          wait: false,
-        });
+        expect(open).toHaveBeenCalledWith(
+          `http://localhost:${port}/index.html`,
+          {
+            app: {
+              name: 'google-chrome',
+              arguments: ['--incognito', '--new-window'],
+            },
+            wait: false,
+          }
+        );
         expect(loggerWarnSpy).toHaveBeenCalledWith(
-          'Unable to open "http://localhost:8117/index.html" page in "google-chrome" app with "--incognito --new-window" arguments. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".'
+          `Unable to open "http://localhost:${port}/index.html" page in "google-chrome" app with "--incognito --new-window" arguments. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".`
         );
 
         loggerWarnSpy.mockRestore();
@@ -820,7 +849,7 @@ describe('"open" option', () => {
     const compiler = webpack(config);
     const server = createServer(compiler, {
       open: {
-        target: ['first.html', 'http://localhost:8117/second.html'],
+        target: ['first.html', `http://localhost:${port}/second.html`],
         app: {
           name: 'google-chrome',
           arguments: ['--incognito', '--new-window'],
@@ -835,7 +864,7 @@ describe('"open" option', () => {
       server.close(() => {
         expect(open).toHaveBeenNthCalledWith(
           1,
-          'http://localhost:8117/first.html',
+          `http://localhost:${port}/first.html`,
           {
             wait: false,
             app: {
@@ -846,7 +875,7 @@ describe('"open" option', () => {
         );
         expect(open).toHaveBeenNthCalledWith(
           2,
-          'http://localhost:8117/second.html',
+          `http://localhost:${port}/second.html`,
           {
             wait: false,
             app: {
@@ -857,11 +886,11 @@ describe('"open" option', () => {
         );
         expect(loggerWarnSpy).toHaveBeenNthCalledWith(
           1,
-          'Unable to open "http://localhost:8117/first.html" page in "google-chrome" app with "--incognito --new-window" arguments. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".'
+          `Unable to open "http://localhost:${port}/first.html" page in "google-chrome" app with "--incognito --new-window" arguments. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".`
         );
         expect(loggerWarnSpy).toHaveBeenNthCalledWith(
           2,
-          'Unable to open "http://localhost:8117/second.html" page in "google-chrome" app with "--incognito --new-window" arguments. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".'
+          `Unable to open "http://localhost:${port}/second.html" page in "google-chrome" app with "--incognito --new-window" arguments. If you are running in a headless environment, please do not use the "open" option or related flags like "--open", "--open-target", and "--open-app".`
         );
 
         loggerWarnSpy.mockRestore();
